DŨng 12/4

+++ WebCafe/Areas/Admin/Controllers/SanPhamsController.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -19,13 +19,11 @@ namespace WebCafe.Areas.Admin.Controllers
             _context = context;
         }
 
-
         public async Task<IActionResult> Index()
         {
             var lstSanPham = await _context.SanPhams.Include(s => s.MaDmNavigation).ToListAsync();
             return View(lstSanPham);
         }
-
 
         public async Task<IActionResult> Details(int? id)
         {
@@ -45,13 +43,12 @@ namespace WebCafe.Areas.Admin.Controllers
             return View(sanPham);
         }
 
-
         public IActionResult Create()
         {
+            ViewBag.ThongBao = "";
             ViewData["MaDm"] = new SelectList(_context.DanhMucSps, "MaDm", "TenDm");
             return View();
         }
-
 
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -59,14 +56,20 @@ namespace WebCafe.Areas.Admin.Controllers
         {
             if (ModelState.IsValid)
             {
-                _context.Add(sanPham);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                // Kiểm tra sản phẩm đã tồn tại chưa?
+                if (!SanPhamExists(sanPham.MaDm, sanPham.TenSp))
+                {
+                    sanPham.NgaySua = sanPham.CreateDate;
+                    _context.Add(sanPham);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                // Đưa ra thông báo sản phẩm đã tồn tại
+                ViewBag.ThongBao = "Sản phẩm đã tồn tại, thêm mới không thành công";
             }
             ViewData["MaDm"] = new SelectList(_context.DanhMucSps, "MaDm", "TenDm", sanPham.MaDm);
             return View(sanPham);
         }
-
 
         public async Task<IActionResult> Edit(int? id)
         {
@@ -74,45 +77,31 @@ namespace WebCafe.Areas.Admin.Controllers
             {
                 return NotFound();
             }
-
             var sanPham = await _context.SanPhams.FindAsync(id);
             if (sanPham == null)
             {
                 return NotFound();
             }
+            ViewBag.ThongBao = "";
             ViewData["MaDm"] = new SelectList(_context.DanhMucSps, "MaDm", "TenDm", sanPham.MaDm);
             return View(sanPham);
         }
 
-
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("MaSp,MaDm,TenSp,AnhSp,VideoSp,GiaSp,TrangThai,SoLuong,BestSeller,CreateDate,NgaySua,MotaSp")] SanPham sanPham)
+        public async Task<IActionResult> Edit([Bind("MaSp,MaDm,TenSp,AnhSp,VideoSp,GiaSp,TrangThai," +
+            "SoLuong,BestSeller,CreateDate,NgaySua,MotaSp")] SanPham sanPham)
         {
-            if (id != sanPham.MaSp)
-            {
-                return NotFound();
-            }
-
             if (ModelState.IsValid)
             {
-                try
+                if (!SanPhamExists(sanPham.MaDm, sanPham.TenSp))
                 {
                     _context.Update(sanPham);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!SanPhamExists(sanPham.MaSp))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                // Đưa ra thông báo sản phẩm đã tồn tại
+                ViewBag.ThongBao = "Sản phẩm đã tồn tại, sửa sản phẩm không thành công";
             }
             ViewData["MaDm"] = new SelectList(_context.DanhMucSps, "MaDm", "TenDm", sanPham.MaDm);
             return View(sanPham);
@@ -125,30 +114,39 @@ namespace WebCafe.Areas.Admin.Controllers
                 return NotFound();
             }
 
-            var sanPham = await _context.SanPhams
-                .Include(s => s.MaDmNavigation)
-                .FirstOrDefaultAsync(m => m.MaSp == id);
+            var sanPham = await _context.SanPhams.FindAsync(id);
             if (sanPham == null)
             {
                 return NotFound();
             }
-
+            ViewBag.ThongBao = "";
             return View(sanPham);
         }
 
-        [HttpPost, ActionName("Delete")]
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> Delete(int id)
         {
             var sanPham = await _context.SanPhams.FindAsync(id);
-            _context.SanPhams.Remove(sanPham);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (!SanPhamTrongChiTietDonHang(id) && sanPham!=null)
+            {
+                _context.SanPhams.Remove(sanPham);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            // Đưa ra thông báo sản phẩm đã tồn tại
+            ViewBag.ThongBao = "Sản phẩm đã được bán, không được phép xóa";
+            return View(sanPham);
         }
 
-        private bool SanPhamExists(int id)
+        private bool SanPhamExists(int MaDm, string TenSp)
         {
-            return _context.SanPhams.Any(e => e.MaSp == id);
+            return _context.SanPhams.Any(e => e.MaDm == MaDm && e.TenSp == TenSp);
+        }
+
+        public bool SanPhamTrongChiTietDonHang(int MaSp)
+        {
+            return _context.ChiTietDonHangs.Any(x => x.MaSp == MaSp);
         }
     }
 }
